+++ Sources/HandVector/HVJointFingerHand.swift
@@ -1,0 +1,48 @@
+//
+//  HVJointJsonModel.swift
+//
+//
+//  Created by xu on 2024/8/16.
+//
+
+import ARKit
+
+protocol HVJoint {
+    var name: HandSkeleton.JointName.NameCodingKey { get }
+    var isTracked: Bool { get }
+    var transform: simd_float4x4 { get }
+}
+
+
+public enum HVJointOfFinger: CaseIterable {
+    case thump
+    case indexFinger
+    case middleFinger
+    case ringFinger
+    case littleFinger
+    case wristMetacarpal
+    case foreArm
+    
+    public var jointGroupNames: [HandSkeleton.JointName] {
+        switch self {
+        case .thump:
+            return [.thumbKnuckle, .thumbIntermediateBase, .thumbIntermediateTip, .thumbTip]
+        case .indexFinger:
+            return [.indexFingerKnuckle, .indexFingerIntermediateBase, .indexFingerIntermediateTip, .indexFingerTip]
+        case .middleFinger:
+            return [.middleFingerKnuckle, .middleFingerIntermediateBase, .middleFingerIntermediateTip, .middleFingerTip]
+        case .ringFinger:
+            return [.ringFingerKnuckle, .ringFingerIntermediateBase, .ringFingerIntermediateTip, .ringFingerTip]
+        case .littleFinger:
+            return [.littleFingerKnuckle, .littleFingerIntermediateBase, .littleFingerIntermediateTip, .littleFingerTip]
+        case .wristMetacarpal:
+            return [.wrist, .indexFingerMetacarpal, .middleFingerMetacarpal, .ringFingerMetacarpal, .littleFingerMetacarpal]
+        case .foreArm:
+            return [.forearmWrist, .forearmArm]
+        }
+    }
+    
+    public static let allFingers: Set<HVJointOfFinger> = [.thump, .indexFinger, .middleFinger, .ringFinger, .littleFinger]
+    public static let allFingersAndWrist: Set<HVJointOfFinger> = [.thump, .indexFinger, .middleFinger, .ringFinger, .littleFinger, .wristMetacarpal]
+    public static let allFingersAndWristAndForearm: Set<HVJointOfFinger> = [.thump, .indexFinger, .middleFinger, .ringFinger, .littleFinger, .wristMetacarpal, .foreArm]
+}
